Move ParticleProductionFilter into its own source file
